% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{checkCorrelation}
\alias{checkCorrelation}
\title{Check correlation matrix}
\usage{
checkCorrelation(m, returnMessage = FALSE, na.allowed = TRUE)
}
\arguments{
\item{m}{TBA}

\item{returnMessage}{TBA}

\item{na.allowed}{TBA}
}
\value{
Logical
}
\description{
Sanity checks for the correlation matrix.
}
\details{
Checks the following properties:
\itemize{
  \item Values must be between -1 and 1.
  \item Diagonal must be equal to 1.
  \item Matrix must be symmetric.
}
}
\examples{
NULL
}
