% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gMCPLite-package.R
\docType{package}
\name{gMCPLite-package}
\alias{gMCPLite}
\alias{gMCPLite-package}
\title{gMCPLite: Lightweight Graph Based Multiple Comparison Procedures}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

A lightweight fork of 'gMCP' with functions for graphical described multiple test procedures introduced in Bretz et al. (2009) \doi{10.1002/sim.3495} and Bretz et al. (2011) \doi{10.1002/bimj.201000239}. Implements a flexible function using 'ggplot2' to create multiplicity graph visualizations. Contains instructions of multiplicity graph and graphical testing for group sequential design, described in Maurer and Bretz (2013) \doi{10.1080/19466315.2013.807748}, with necessary unit testing using 'testthat'.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://merck.github.io/gMCPLite/}
  \item \url{https://github.com/Merck/gMCPLite}
  \item Report bugs at \url{https://github.com/Merck/gMCPLite/issues}
}

}
\author{
\strong{Maintainer}: Yalin Zhu \email{yalin.zhu@merck.com} (\href{https://orcid.org/0000-0003-3830-8660}{ORCID})

Authors:
\itemize{
  \item Yilong Zhang \email{elong0527@gmail.com}
  \item Xuan Deng \email{xuan.deng@merck.com}
  \item Keaven Anderson \email{keaven_anderson@merck.com}
  \item Nan Xiao \email{nan.xiao1@merck.com} (\href{https://orcid.org/0000-0002-0250-5673}{ORCID})
}

Other contributors:
\itemize{
  \item Kornelius Rohmeyer \email{rohmeyer@small-projects.de} (gMCP author) [contributor]
  \item Florian Klinglmueller \email{float@lefant.net} (gMCP author) [contributor]
  \item gMCP project contributors (gMCP package) [copyright holder]
  \item Merck & Co., Inc., Rahway, NJ, USA and its affiliates [copyright holder]
}

}
\keyword{internal}
