% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{permutations}
\alias{permutations}
\title{Permutation for a design matrix}
\usage{
permutations(n)
}
\arguments{
\item{n}{dimension of the matrix}
}
\value{
a n*(2^n) dimensional matrix
}
\description{
Permutation for a design matrix
}
\examples{
permutations(3)

}
